Split out column logic to redux
