+++ src/actions/ActionTypes.js
@@ -12,11 +12,16 @@
 'use strict';
 
 const ActionTypes = {
+  COLUMN_REORDER: 'COLUMN_REORDER',
+  COLUMN_REORDER_END: 'COLUMN_REORDER_END',
+  COLUMN_REORDER_MOVE: 'COLUMN_REORDER_MOVE',
+  COLUMN_RESIZE: 'COLUMN_RESIZE',
   PROP_CHANGE: 'PROP_CHANGE',
   SCROLL_BY: 'SCROLL_BY',
   SCROLL_END: 'SCROLL_END',
   SCROLL_START: 'SCROLL_START',
   SCROLL_TO: 'SCROLL_TO',
+  SCROLL_X: 'SCROLL_X',
 };
 
 module.exports = ActionTypes;
